Add basic BlockFactory and remove invalid field from Block json
Adds a BlockFactory class for loading a collection of blocks that
can then be duplicated for use in the workspace. Just a very basic
implementation for now.

Also, there's no 'name' property in the Block json, the id is the
name, so this removes references to name from Block and its builder.